Separate the interface structs in an independent file
